requirements.txt
# Core Libraries
pytube
youtube-transcript-api
requests
python-dotenv
# Transcription
openai-whisper
torch

# Summarization
transformers
sentencepiece

# Orchestration and Agents
langchain
openai
langchain_community
# Web Framework
flask
flask-cors

# Deployment
gunicorn
